pull requset를 위한 파일 경로 수정
