+++ src/main/java/com/JoinUs/dp/common/exception/NotFoundException.java
@@ -1,0 +1,6 @@
+package com.JoinUs.dp.common.exception;
+
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String msg) { super(msg); }
+}
