+++ src/main/java/com/JoinUs/dp/controller/Snippet.java
@@ -1,0 +1,7 @@
+package com.JoinUs.dp.controller;
+
+public class Snippet {
+	public static void main(String[] args) {
+	}
+}
+
