--- src/main/java/com/JoinUs/dp/global/common/exception/GlobalExceptionHandler.java
@@ -1,24 +1,0 @@
-package com.JoinUs.dp.global.common.exception;
-
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import com.JoinUs.dp.global.common.response.Response;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<Response<Void>> handleNotFound(NotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new Response<>(HttpStatus.NOT_FOUND.value(), null, ex.getMessage()));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Response<Void>> handleOther(Exception ex) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(new Response<>(HttpStatus.INTERNAL_SERVER_ERROR.value(), null, "서버 오류가 발생했습니다."));
-    }
-}
