--- src/main/java/com/JoinUs/dp/global/common/exception/NotFoundException.java
@@ -1,6 +1,0 @@
-package com.JoinUs.dp.global.common.exception;
-
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String msg) { super(msg); }
-}
