--- src/test/java/com/JoinUs/dp/DpApplicationTests.java
@@ -1,13 +1,0 @@
-package com.JoinUs.dp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DpApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
